Provide a default value for read/write options
This CL marks FileSystemReadWriteOptions as optional, and sets the 'at'
member default to 0. This means that access handle reads and writes
will start from the beginning of a file if options are not provided.

It also cleans up the affected test by either reusing the
encoder/decoder variables or removing them when they would only be used
once.

Bug: 1218431
Change-Id: Ibbd34a92dd9cdc44cd58c96860698ffddd8eb469