HighlightInheritance: fix implicit defaulting of ‘color’ at :root
Highlight pseudos redefine the inherited value of ‘color’ for :root to
be ‘currentColor’ (which is special for highlights) rather than the
initial value, such that the implicit default and ‘inherit’ and
‘unset’ all mean ‘currentColor’, while ‘initial’ still refers to the
initial value (‘CanvasText’).

The special concept of ‘currentColor’ for highlights is encoded in the
ColorIsCurrentColor extra field, with the help of css_longhand::Color
methods (for explicit defaults) and the default value of the field
itself (for implicit defaulting at :root).

This works if we actually resolve highlight styles at :root, such as
with getComputedStyle(document.documentElement) or if there are rules
for the :root highlight setting other properties. But if the :root
highlight is completely unstyled, the StyleResolverState ctor will
default its parent style to the ElementResolveContext’s parent style,
that is, the originating(!) parent element style.

This patch fixes that “initial-value bug” (CL:3173223), by making the
StyleResolverState ctor leave the parent and layout parent unset when
resolving highlight styles with highlight inheritance.

To keep tests passing, we also need to fix a bug where highlighting a
visited link yields a currentColor foreground when the visited style
isn’t currentColor but the unvisited style is currentColor. This patch
does so by splitting ColorIsCurrentColor (CL:3506351, CL:3563099) into
a pair of flags, one visited and one unvisited.

Fixed: 1341621
Change-Id: I2ac356ef8c69756d5d95cd09e639ad32980d41bd