Fix wpt test for css-generated text on inline element
The wpt test for css-generated text on inline elements
was failing since the reported LCP was being triggered
for the next block-level ancestor as outlined in
https://github.com/WICG/element-timing#text-considerations.

Fixed the test to account for the aggregation of LCP
entries at the next block-level ancestor.

Also prevent the test from timing out if no LCP entries
are found.

Bug: 1321377
Change-Id: I65f4eb9cae42e6d19e485983843d5f49d852423c