Add WPT-based <slot> fallback reference tests
During release M89, a bug was noticed after release and affected users
in production. Postmortem [1] was created to resolve this bug quickly.
This shows holes in the existing slots fallback tests, especially
when it comes to the final rendered output.

To remediate this, we are adding reference tests for slots
fallback that cover most common use cases.

[1] https://docs.google.com/document/d/1IsypzwRE3-MLLFhz6ldaxxIYsn_vLRGPzvv9ftZSoqU/

Change-Id: Ib76d516178ac055591fb51bd04f9be4cd96addc6
Fixed: 1185209